Fixed background noise in the refined mask
As a result of registration it can happen that the refined mask, given as a sum of soft segmentation of wm, gm and csf, has small non-zero values in the background. These can pose an issue for computing 95 % credibility interval when threasholding the soft mask. This problem was fixed by multiplying the soft mask by the coarse brain mask from the rigid registration.